src/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
requests>=2.31.0
# torch and torchvision installed via install_torch.py
pillow>=10.0.0
opencv-python==4.10.0.84
numpy>=1.24.0
timm>=0.9.0
ultralytics==8.3.63
segment-anything-py==1.0
httpx==0.28.1
azure-storage-blob==12.19.0
python-dotenv==1.0.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.36
icrawler==0.6.6
transformers>=4.38.0
nltk>=3.8.1
pytest
httpx